server/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.2
flask_cors
SQLAlchemy
flask_sqlalchemy
flask_dotenv
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
python-jose 
six 